Create Task 3 codsoft
This project is a Python-based password generator that creates secure and random passwords of user-defined length. It combines letters, digits, and special characters to ensure strong security. The program is interactive, validates user input, and allows generating multiple passwords in one run. It is beginner-friendly and demonstrates the use of Python’s random and string modules, while also being extendible for future upgrades like GUI or web integration.